requirements.txt
streamlit
pypdf
llama-index
python-dotenv
pandas

